refactor(protocol): declare AppwireClientLike in the package, not the test double
AppwireClientLike is the connection seam every store, hook and component in
the web app is written against, and it was declared inside FakeClient, a test
double that tsconfig.build.json does not ship. Production code therefore
depended on a test helper for its core interface, and an installed consumer
could not name the type at all.

The interface moves to protocol/clientLike.ts, joins the package build and the
entry point, and fakeClient.ts imports it like everyone else. client.test.ts
now assigns a real AppwireClient to the interface: its eight methods are
declared by lookup on the class, but `state` and `terminalReason` are spelled
out, so only that assignment catches an accessor drifting away from the seam.
