+++ cmd/evener-hub/frontend/src/panes/spawn/useSpawnSlashCatalog.ts
@@ -1,6 +1,6 @@
 import { useCallback, useEffect, useRef, useState } from "react";
+import type { AppwireClientLike } from "../../protocol/clientLike";
 import { errorText } from "../../protocol/errors";
-import type { AppwireClientLike } from "../../protocol/testing/fakeClient";
 import type { LaunchConfigLayer, SpawnSlashCatalogResponse } from "../../protocol/types.gen";
 
 export const SPAWN_SLASH_CATALOG_DEBOUNCE_MS = 250;
