+++ cmd/evener-hub/frontend/src/protocol/index.ts
@@ -30,6 +30,7 @@ export type { AskAnswerItem, AskResolution } from "./askAnswers";
 export { composeAskAnswers } from "./askAnswers";
 export type { AnyNotification, AppwireClientOptions, ConnectionState, TerminalReason } from "./client";
 export { APPWIRE_PROTOCOL_VERSION, AppwireClient } from "./client";
+export type { AppwireClientLike } from "./clientLike";
 export { ConnectionClosedError, RequestTimeoutError, WireError } from "./errors";
 export type { JobLogTail } from "./jobOutput";
 export { parseJobLogTail } from "./jobOutput";
