+++ cmd/evener-hub/frontend/src/protocol/scripts/qualify-package.mjs
@@ -32,6 +32,7 @@ async function qualify() {
   const shippedModules = [
     "index",
     "client",
+    "clientLike",
     "errors",
     "transport",
     "types.gen",
@@ -96,6 +97,7 @@ async function qualify() {
   // check covers each module's packed .d.ts and not just its runtime half.
   const typeExports = [
     "AppwireClientOptions",
+    "AppwireClientLike",
     "WebSocketLike",
     "AskAnswerItem",
     "ActivityTree",
