+++ cmd/evener-hub/frontend/src/protocol/tsconfig.build.json
@@ -14,6 +14,7 @@
   "files": [
     "index.ts",
     "client.ts",
+    "clientLike.ts",
     "errors.ts",
     "transport.ts",
     "types.gen.ts",
