+++ cmd/evener-hub/frontend/src/shell/clientContext.tsx
@@ -7,7 +7,7 @@
 // seam for stores that have no connect() path of their own (see
 // stores/connection.ts).
 import { createContext, type ReactNode, useContext } from "react";
-import type { AppwireClientLike } from "../protocol/testing/fakeClient";
+import type { AppwireClientLike } from "../protocol/clientLike";
 
 const ClientContext = createContext<AppwireClientLike | null>(null);
