+++ cmd/evener-hub/frontend/src/shell/rail/actions.ts
@@ -3,8 +3,8 @@
 // deletion use the typed AppWire client. No optimistic UI: callers await the
 // response's exact navigation targets before removing their overlay.
 
+import type { AppwireClientLike } from "../../protocol/clientLike";
 import { WireError } from "../../protocol/errors";
-import type { AppwireClientLike } from "../../protocol/testing/fakeClient";
 import type {
   FavoriteSetResponse,
   NavigationMutation,
