+++ cmd/evener-hub/frontend/src/stores/launchConfig.ts
@@ -14,7 +14,7 @@
 // cwd+layer, so there is no single value to memoize the way schema's is.
 import { useStore } from "zustand";
 import { createStore } from "zustand/vanilla";
-import type { AppwireClientLike } from "../protocol/testing/fakeClient";
+import type { AppwireClientLike } from "../protocol/clientLike";
 import type {
   LaunchConfigLayer,
   LaunchConfigResolved,
