+++ cmd/evener-hub/frontend/src/stores/settingsOverview.ts
@@ -24,8 +24,8 @@
 
 import { useStore } from "zustand";
 import { createStore } from "zustand/vanilla";
+import type { AppwireClientLike } from "../protocol/clientLike";
 import { errorText } from "../protocol/errors";
-import type { AppwireClientLike } from "../protocol/testing/fakeClient";
 import type { SettingsOverviewResponse } from "../protocol/types.gen";
 import { connectionStore } from "./connection";
