+++ cmd/evener-hub/frontend/src/stores/transcriptDisplay.ts
@@ -1,8 +1,8 @@
 import { useStore } from "zustand";
 import { createStore, type StoreApi } from "zustand/vanilla";
 import { transitionTranscriptViews } from "../panes/session/transcript/flow/transcriptViewRegistry";
+import type { AppwireClientLike } from "../protocol/clientLike";
 import { WireError } from "../protocol/errors";
-import type { AppwireClientLike } from "../protocol/testing/fakeClient";
 import type { AnyNotification } from "../protocol/types.gen";
 import { isMobileViewport, subscribeMobileViewport } from "../shell/useIsMobile";
 import {
